refactor(renderer): extract App helpers, auth, and catalog hooks
Move pure direct-launch and ownership mutators into lib modules, pull
auth/session login flows into useAuthSession, and catalog loading into
useCatalogData. Stream launch/claim/recovery stays in App; only pure
stream helpers were extracted pending a safer session-runtime split.
